Bug Joining Lobbies (really weird)
The value of "i" when adding onClick listeners to the join buttons in LobbyManager.cs somehow changes by the time the button is clicked.

I think that it is something to do with memory since I is used in a for loop and it doesn't pass the value of the int i, but a ref somehow.

Will be fixed by adding a temp integer variable to literally just do int temp = i; then passing in temp instead but wanted to save this bug to a commit bc its interesting.