Rubber Banding discovery
The inputs aren't getting lost, the server is somehow skipping some ticks (about 3.5%) of them.

This is shown by running "beginILM"/"endILM" in the console.

It compares tick numbers being run by the server against the client and shows that the client is running more frames than the server somehow.